test: cast floats to int for numpy slicing

+++ test/test_autochisq.py
@@ -82,7 +82,7 @@ class TestAutochisquare(unittest.TestCase):
 
         # generate sin-gaussian signal
         time = np.arange(0, len(hp))*self.del_t
-        Nby2 = len(hp)/2
+        Nby2 = int(len(hp)/2)
         sngt = np.zeros(len(hp))
         for i in range(len(hp)):
             sngt[i] = 9.0e-21*exp(-(time[i]-time[Nby2])**2/self.Q)*sin(self.om*time[i])
